Pathetic workaround to show exception text.
I have no idea what is going on with this. At some point in the past I was
seeing error text fine, but now exceptions (from NameErrors to assertEquals()
failures) are showing an abbreviated, useless message. This restores the
message, but at the expense of an extra indentation through the whole code.

I was in such a bad mood about it I decided to only add a single space for the
indent. Maybe I'll change my mind and add the remaining three, I don't know.